Update RegisterForm.cs -- To Implement "System - User Verification"
It checks to see: 
when a student is registering for the course scheduler, that the student is already in the "WSU System".

My concern at the moment is: 
for this validation system to work, the student's email has to be already in the database. There is no way for the student to do this, it has to be brute-forced written into the program.